refactor(DEV-3834): flatten div elements and map author to dcterms:creator for Google AI Pilot and move old feeds to xml.js
